refactor styles to match menu schema.
